refactor(EmptyState): combine header and footer rules together
