+++ packages/eslint-plugin-pf-codemods/index.js
@@ -24,8 +24,7 @@ const warningRules = [
   "conditional-aria",
   "datePicker-warn-appendTo-default-value-changed",
   "datepicker-warn-helperText",
-  "emptyState-warn-add-footer",
-  "emptyState-warn-add-header",
+  "emptyState-warn-change-structure",
   "horizontalSubnav-warn-ariaLabel",
   "label-warn-truncated-default",
   "nav-warn-flyouts-now-inline",
