+++ packages/eslint-plugin-pf-codemods/test/rules/v5/emptyState-warn-change-structure.js
@@ -1,13 +1,14 @@
 const ruleTester = require("../../ruletester");
-const rule = require("../../../lib/rules/v5/emptyState-warn-add-footer");
+const rule = require("../../../lib/rules/v5/emptyState-warn-change-structure");
 
-ruleTester.run("emptyState-warn-add-footer", rule, {
+ruleTester.run("emptyState-warn-change-structure", rule, {
   valid: [
     {
-      code: `import { Button, EmptyState, EmptyStateBody, EmptyStateActions, EmptyStateFooter } from '@patternfly/react-core';
+      code: `import { Button, EmptyState, EmptyStateActions, EmptyStateBody, EmptyStateIcon, Title, EmptyStateHeader, EmptyStateFooter } from '@patternfly/react-core'; 
       <EmptyState>
+        <EmptyStateHeader titleText="Empty state" icon={<EmptyStateIcon icon={CubesIcon} />} headingLevel="h5" />
         <EmptyStateBody>
-          Some content
+          Some other content.
         </EmptyStateBody><EmptyStateFooter>
         <EmptyStateActions>
           <Button variant="primary">Primary action</Button>
@@ -22,8 +23,12 @@ ruleTester.run("emptyState-warn-add-footer", rule, {
     {
       // No @patternfly/react-core import
       code: `<EmptyState>
+        <EmptyStateIcon icon={CubesIcon} />
+        <Title headingLevel="h5" size="4xl">
+          Empty state
+        </Title>
         <EmptyStateBody>
-          Some content
+          Some other content.
         </EmptyStateBody>
         <EmptyStateActions>
           <Button variant="primary">Primary action</Button>
@@ -38,10 +43,14 @@ ruleTester.run("emptyState-warn-add-footer", rule, {
   ],
   invalid: [
     {
-      code: `import { Button, EmptyState, EmptyStateBody, EmptyStateActions } from '@patternfly/react-core';
+      code: `import { Button, EmptyState, EmptyStateActions, EmptyStateBody, EmptyStateIcon, Title } from '@patternfly/react-core'; 
       <EmptyState>
+        <EmptyStateIcon icon={CubesIcon} />
+        <Title headingLevel="h5" size="4xl">
+          Empty state
+        </Title>
         <EmptyStateBody>
-          Some content
+          Some other content.
         </EmptyStateBody>
         <EmptyStateActions>
           <Button variant="primary">Primary action</Button>
@@ -52,10 +61,11 @@ ruleTester.run("emptyState-warn-add-footer", rule, {
           <Button variant="link">Here</Button>
         </EmptyStateActions>
       </EmptyState>`,
-      output: `import { Button, EmptyState, EmptyStateBody, EmptyStateActions, EmptyStateFooter } from '@patternfly/react-core';
+      output: `import { Button, EmptyState, EmptyStateActions, EmptyStateBody, EmptyStateIcon, Title, EmptyStateHeader, EmptyStateFooter } from '@patternfly/react-core'; 
       <EmptyState>
+        <EmptyStateHeader titleText="Empty state" icon={<EmptyStateIcon icon={CubesIcon} />} headingLevel="h5" />
         <EmptyStateBody>
-          Some content
+          Some other content.
         </EmptyStateBody><EmptyStateFooter>
         <EmptyStateActions>
           <Button variant="primary">Primary action</Button>
@@ -68,8 +78,12 @@ ruleTester.run("emptyState-warn-add-footer", rule, {
       </EmptyStateFooter></EmptyState>`,
       errors: [
         {
-          message: `add missing imports EmptyStateFooter from @patternfly/react-core`,
+          message: `add missing imports EmptyStateHeader, EmptyStateFooter from @patternfly/react-core`,
           type: "ImportDeclaration",
+        },
+        {
+          message: `EmptyStateHeader component should be added instead of Title and EmptyStateIcon`,
+          type: "JSXElement",
         },
         {
           message: `EmptyStateFooter component should be added to wrap content after EmptyStateBody`,
